requirements.txt
flask>=3.0
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
gunicorn>=21.2
